pyproject.toml
[tool.poetry]
name = "icgl"
version = "0.1.0"
description = "ICGL - Iterative Co-Governance Loop: A Governance-First Intelligence System"
authors = ["Bakheet <bakheet@gmail.com>"]
readme = "README.md"
packages = [{include = "icgl", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
rich = "^13.7.0"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
openai = "^1.12.0"
qdrant-client = "^1.7.0"
pydantic = "^2.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
icgl = "icgl.main:main"
